refactor# add view library and remove common lib
