+++ library-view/.gitignore
