+++ library-view/src/androidTest/java/life/chenshi/keepaccounts/library/view/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package life.chenhsi.commonlib
+package life.chenshi.keepaccounts.library.view
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("life.chenhsi.commonlib.test", appContext.packageName)
+        assertEquals("life.chenshi.keepaccounts.library.view.test", appContext.packageName)
     }
 }
